+++ projects/BetBot/src/sagas/framework/saga.ts
@@ -1,0 +1,34 @@
+import { TaskError } from './error';
+import { Task } from './task';
+
+/* eslint-disable @typescript-eslint/ban-types */
+interface ITaskMap {
+  [name: string]: Function;
+}
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export class Saga {
+  taskMap: ITaskMap = {};
+
+  addTask(taskName: string, func: Function): void {
+    this.taskMap[taskName] = func;
+  }
+
+  async start(task: Task): Promise<void> {
+    if (!task) return;
+    // eslint-disable-next-line no-prototype-builtins
+    if (!this.taskMap.hasOwnProperty(task.name))
+      throw Error(`TASK NOT SUPPORTED: ${task.name}`);
+
+    try {
+      task.setOutput(await this.taskMap[task.name](task.taskInfo.input));
+      task.completeTask('pass');
+    } catch (err) {
+      if (err instanceof TaskError) task.setOutput(err.getTaskInfo());
+      task.completeTask('fail');
+    }
+
+    const nextTask = task.nextTask();
+    if (nextTask) this.start(nextTask);
+  }
+}
