+++ projects/BetBot/src/sagas/framework/task.ts
@@ -1,0 +1,94 @@
+import { ChatInputCommandInteraction } from 'discord.js';
+
+/* eslint-disable @typescript-eslint/no-explicit-any */
+export interface ITaskData {
+  interaction?: ChatInputCommandInteraction;
+  [name: string]: any;
+}
+
+interface ITaskHistory {
+  name: string;
+  status: 'pass' | 'fail';
+  input: ITaskData;
+  output: ITaskData;
+}
+
+interface ITaskInfo {
+  input: ITaskData;
+  output: ITaskData;
+  history: ITaskHistory[];
+}
+
+// eslint-disable-next-line @typescript-eslint/no-unused-vars
+export class Task {
+  name: string;
+
+  private passTask: Task;
+
+  private failTask: Task;
+
+  private status: 'pass' | 'fail';
+
+  taskInfo: ITaskInfo = {
+    input: undefined,
+    output: undefined,
+    history: [],
+  };
+
+  constructor(name: string) {
+    this.name = name;
+  }
+
+  completeTask(status: 'pass' | 'fail'): void {
+    this.status = status;
+  }
+
+  pass(passTask: Task): void {
+    this.passTask = passTask;
+  }
+
+  fail(failTask: Task): void {
+    this.failTask = failTask;
+  }
+
+  setInput(input: ITaskData): void {
+    this.taskInfo.input = input;
+  }
+
+  setOutput(output: ITaskData): void {
+    this.taskInfo.output = output;
+  }
+
+  private taskToHistory(): ITaskHistory {
+    return {
+      name: this.name,
+      status: this.status,
+      input: this.taskInfo.input,
+      output: this.taskInfo.output,
+    };
+  }
+
+  addHistory(task: Task): void {
+    this.taskInfo.history = task.taskInfo.history;
+    this.taskInfo.history.push(task.taskToHistory());
+  }
+
+  nextTask(): Task | undefined {
+    // FAILURE TASK
+    if (this.status === 'fail') {
+      if (!this.failTask) return undefined;
+      this.failTask.setInput(this.taskInfo.output);
+      this.failTask.addHistory(this);
+      return this.failTask;
+    }
+
+    // SUCCESS TASK
+    if (this.status === 'pass') {
+      if (!this.passTask) return undefined;
+      this.passTask.setInput(this.taskInfo.output); // SET THE INPUT TO THE NEXT TASK AS THE OUTPUT FROM PREVIOUS
+      this.passTask.addHistory(this);
+      return this.passTask;
+    }
+    throw Error(`TASK INCOMPLETE: ${this.name}`);
+  }
+}
